requirements.txt
torch>=2.0.0
transformers>=4.37.0
accelerate>=0.25.0
Pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-image>=0.21.0
opencv-python>=4.8.0
matplotlib>=3.7.0
tqdm>=4.65.0
pyyaml>=6.0
huggingface-hub>=0.20.0
peft>=0.7.0
sentencepiece>=0.1.99
protobuf>=3.20.0
